# server/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/semester/{semester}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "semester",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/name/{name}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "semester",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/student/{studentId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "semester",
            "in": "query",
            "type": "string"
          },
          {
            "name": "weekday",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/student/{studentId}/incoming": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "semester",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/semester/{semester}/start": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "semester",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/tx/{txHash}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "txHash",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}